Credit the correct authors for the IF1 disassembly
While I appreciate the acknowledgement, the only credit I can take for the IF1
disassembly is for hosting it on my server. I had no part in creating it.

The original page where this disassembly was hosted listed Geoff Wearmouth as
the maintainer. There is also a comment at the end of the linked file that
credits others with specific parts of the documentation. It is them who should
get the credit for their excellent work.

I briefly explain how I ended up hosting the files here:

https://www.tablix.org/~avian/spectrum/rom/